requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
pytesseract==0.3.10
Pillow==10.0.0
gunicorn==21.2.0
cloudinary==1.36.0
requests==2.31.0
PyMuPDF==1.23.8
easyocr==1.7.0